requirements.txt
filetype~=1.2.0
uvicorn~=0.28.0
fastapi~=0.110.0
pydub~=0.25.1
python-multipart~=0.0.9
openai-whisper
